tft and inverse tft for 32 and 64 bit

+++ flint-extras/nmod_extra/src/nmod_extra.c
@@ -12,7 +12,7 @@
 int next_power_of_two(mp_limb_t x)
 {
     int i = 0, j = 1;
-    while (j < x)
+    while (j < (int) x)
     {
         i++;
         j <<= 1;
